requirements.txt
Flask==0.12.2
Werkzeug<0.13.0,>=0.12.0
flask_wtf>0.12
flask-sqlalchemy
pymysql
google-auth
requests
